Move Models and DBInitializer -> Infrastructure
