+++ VerbTrainer/Infrastructure/Data/Models/ErrorViewModel.cs
