+++ VerbTrainer/Infrastructure/Data/Models/Hebrew/Conjugation.cs
@@ -3,7 +3,7 @@ using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
 
-namespace VerbTrainer.Models.Domain
+namespace VerbTrainer.Infrastructure.Data.Models.Hebrew
 {
     public class Conjugation
     {
