+++ VerbTrainer/Infrastructure/Data/Models/Hebrew/DeckVerb.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace VerbTrainer.Models.Domain
+namespace VerbTrainer.Infrastructure.Data.Models.Hebrew
 {
 	public class DeckVerb
 	{
