+++ VerbTrainer/VerbTrainer.csproj
@@ -26,26 +26,31 @@
   </ItemGroup>
   <ItemGroup>
     <None Remove="allverbs" />
-    <None Remove="Data\" />
-    <None Remove="Models\Domain\" />
     <None Remove="ViewModels\" />
     <None Remove="DTOs\" />
     <None Remove="Infrastructure\" />
     <None Remove="Infrastructure\Messaging\" />
     <None Remove="Infrastructure\Messaging\Producer\" />
     <None Remove="Infrastructure\Messaging\Configuration\" />
+    <None Remove="Infrastructure\Data\" />
+    <None Remove="Infrastructure\DbInitializer\" />
+    <None Remove="Infrastructure\Data\Models\" />
+    <None Remove="Migrations\" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="allverbs" />
   </ItemGroup>
   <ItemGroup>
-    <Folder Include="Data\" />
-    <Folder Include="Models\Domain\" />
     <Folder Include="ViewModels\" />
     <Folder Include="DTOs\" />
     <Folder Include="Infrastructure\" />
     <Folder Include="Infrastructure\Messaging\" />
     <Folder Include="Infrastructure\Messaging\Producer\" />
     <Folder Include="Infrastructure\Messaging\Configuration\" />
+    <Folder Include="Infrastructure\" />
+    <Folder Include="Infrastructure\Data\" />
+    <Folder Include="Infrastructure\DbInitializer\" />
+    <Folder Include="Infrastructure\Data\Models\" />
+    <Folder Include="Migrations\" />
   </ItemGroup>
 </Project>
